[PIPRES-468] renamed services
